requirements.txt
pandas
eventregistry-python
python-dotenv
tqdm
google-generativeai
boto3
url-normalize
requests
rank-bm25
openai
pydantic